+++ tests/Restart/RedisRestartTest.php
@@ -10,14 +10,14 @@ use Tomaj\Hermes\Restart\RedisRestart;
 class RedisRestartTest extends PHPUnit_Framework_TestCase
 {
 
-    public function testNoRedis()
+    public function testInitWithoutRedis()
     {
         $redis = null;
         $this->setExpectedException('InvalidArgumentException', 'Predis\Client or Redis instance required');
         new RedisRestart($redis);
     }
 
-    public function testNoEntry()
+    public function testShouldRestartWithoutRedisEntry()
     {
         $redis = $this->getMock('\Redis', ['get']);
         $redis->expects($this->once())
@@ -28,7 +28,7 @@ class RedisRestartTest extends PHPUnit_Framework_TestCase
         $this->assertFalse($redisRestart->shouldRestart(new \DateTime()));
     }
 
-    public function testFutureEntry()
+    public function testShouldRestartWithFutureEntry()
     {
         $futureTime = (new \DateTime())->modify('+1 month')->format('U');
         $redis = $this->getMock('\Redis', ['get']);
@@ -41,7 +41,7 @@ class RedisRestartTest extends PHPUnit_Framework_TestCase
         $this->assertFalse($redisRestart->shouldRestart(new \DateTime()));
     }
 
-    public function testEntryAfterStartTime()
+    public function testShouldRestartWithEntryAfterStartTime()
     {
         $pastTime = (new \DateTime())->modify('-1 month')->format('U');
         $redis = $this->getMock('\Redis', ['get']);
@@ -54,7 +54,7 @@ class RedisRestartTest extends PHPUnit_Framework_TestCase
         $this->assertFalse($redisRestart->shouldRestart(new \DateTime()));
     }
 
-    public function testSuccess()
+    public function testShouldRestartSuccess()
     {
         $startTime = (new \DateTime())->modify('-1 hour');
         $restartTime = (new \DateTime())->modify('-5 minutes')->format('U');
@@ -66,5 +66,18 @@ class RedisRestartTest extends PHPUnit_Framework_TestCase
 
         $redisRestart = new RedisRestart($redis);
         $this->assertTrue($redisRestart->shouldRestart($startTime));
+    }
+
+    public function testRestartStoredCorrectValueToRedis()
+    {
+        $restartTime = (new \DateTime())->modify('-5 minutes');
+        $redis = $this->getMock('\Redis', ['set', 'get']);
+        $redis->expects($this->once())
+            ->method('set')
+            ->with('hermes_restart', $restartTime->format('U'))
+            ->willReturn(true);
+
+        $redisRestart = new RedisRestart($redis);
+        $this->assertTrue($redisRestart->restart($restartTime));
     }
 }